// manager_frontend/src/client/models/ManagerEquipmentComponentItemResponse.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type ManagerEquipmentComponentItemResponse = {
    id: number;
    equipment_id: number;
    catalog_product_id?: (number | null);
    supplier_id?: (number | null);
    component_type?: string;
    title?: (string | null);
    brand?: (string | null);
    model?: (string | null);
    serial?: (string | null);
    inventory_number?: (string | null);
    supplier_invoice_number?: (string | null);
    supplier_invoice_date?: (string | null);
    notes?: (string | null);
    is_archived?: boolean;
    created_at: string;
    updated_at?: (string | null);
};